+++ CHANGELOG.md
@@ -1,1 +1,8 @@
 
+Version 1.1
+• Now uses const and let to define variables instead of var
+• Removed 'processSubmitButton' 
+• Added 'changeStyle' to replace 'processSubmitButton' so that it can be used on more elements
+• Invalid passwords will now outline the input area in red
+• Simplified some code
+• Added '$' to list of banned characters